Tons more configuration page stuff, bug fixes

+++ src/main/java/duelistmod/abstracts/TokenCard.java
@@ -1,10 +1,12 @@
 package duelistmod.abstracts;
 
+import com.evacipated.cardcrawl.mod.stslib.fields.cards.AbstractCard.CommonKeywordIconsField;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import com.megacrit.cardcrawl.vfx.cardManip.ExhaustCardEffect;
 
+import duelistmod.DuelistMod;
 import duelistmod.actions.unique.PurgeSpecificCard;
 import duelistmod.relics.*;
 
@@ -13,13 +15,16 @@ public class TokenCard extends DuelistCard
 	public TokenCard(String ID, String NAME, String IMG, int COST, String DESCRIPTION, CardType TYPE, CardColor COLOR,CardRarity RARITY, CardTarget TARGET) 
 	{
 		super(ID, NAME, IMG, COST, DESCRIPTION, TYPE, COLOR, RARITY, TARGET);
+		CommonKeywordIconsField.useIcons.set(this, false);
 	}
 
 	@Override
 	public void triggerOnEndOfPlayerTurn()
 	{
-		AbstractDungeon.effectList.add(new ExhaustCardEffect(this));
-		AbstractDungeon.actionManager.addToTop(new PurgeSpecificCard(this, AbstractDungeon.player.hand));
+		if (DuelistMod.tokensPurgeAtEndOfTurn) {
+			AbstractDungeon.effectList.add(new ExhaustCardEffect(this));
+			AbstractDungeon.actionManager.addToTop(new PurgeSpecificCard(this, AbstractDungeon.player.hand));
+		}
 	}
 	
 	@Override
